+++ CHANGELOG.md
@@ -28,6 +28,7 @@ FEATURES:
 CHANGES:
 * Renamed & enhanced command _vapp config network_ to _vapp network external_
 * Several options have been revisited (see README.md for details)
+* Sort VDC, Networks, vApps, Catalogs and Catalog Items by name
 
 VARIOUS:
 * Update dependency vcloud-rest v. 0.4.0